api/requirements.txt
fastapi
uvicorn
motor
pymongo
bcrypt
pyjwt
python-multipart
starlette